+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 # ChangeLog - Aliyun OSS SDK for Go
+## 版本号：1.9.2 日期：2018-11-16
+### 变更内容
+- 变更：添加支持设置request Payer的option
+- 变更：添加支持设置checkpoint目录的option
+- 变更：getobjectmeta接口增加options参数，可以支持传入option选项
+- 变更：listobjecs接口增加options参数，可以支持传入option选项
+- 变更：listmultipartuploads接口增加options参数, 可以支持传入option选项
+- 修复：解决调用接口返回出错时，且返回的http body为空时，打印错误消息不包含"request_id"的问题
+- 变更：abortmultipartupload接口增加options参数, 可以支持传入option选项
+- 变更：completemultipartupload接口增加options参数, 可以支持传入option选项
+
+## 版本号：1.9.1 日期：2018-09-17
+### 变更内容
+ - 变更：支持ipv6
+ - 变更：支持修改对象的存储类型
+ - 修复：修改sample中GetBucketReferer方法名拼写错误
+ - 修复：修复NopCloser在close的时候并不释放内存的内存泄漏问题
+ - 变更：增加ProcessObject接口
+ - 修复：修改图片处理接口参数拼写错误导致无法处理的bug
+ - 修复：增加ListUploadedParts接口的options选项
+ - 修复：增加Callback&CallbackVal选项，支持回调使用
+ - 修复：GetObject接口返回Response，支持用户读取crc等返回值
+ - 修复：当以压缩格式返回数据时，GetObject接口不校验crc
 
 ## 版本号：1.9.0 日期：2018-06-15
 ### 变更内容